Move the store into Todo Page because it's not an app state but a page state
